move grok service in service folder
