+++ airtable-back/src/lib/index.ts
@@ -1,2 +1,2 @@
-export { GroqService } from './groq.service';
+export { GroqService } from '../services/groq.service';
 export { tables, testConnection } from './airtable';
