+++ airtable-back/src/services/ai/chat.prompt.ts
