+++ airtable-back/src/services/ai/recipe-generation.prompt.ts
